Add laptop user wifi and routing hints
Students with laptops find that connecting to a Wi-Fi network at home is unintuitive, as the network manager wishes for elevation.  However, it is a relatively straightforward process via settings->Wi-Fi.

This change gives the option of creating a crontab process which will ensure all users can edit the settings in the network manager.
At the moment, it is only a hint, as the process is manual, and I suspect only admins issuing a laptop would want this change rather than it applying to desktops as well.

The reasoning for editing the crontab in this manner is that an update could replace the configuration file.  Using @reboot avoids this and should apply the change in case of a problem.

Another hint has been added to show how static routes can be added via crontab on boot as well; for particularly broken systems.