Add example printers and Firefox ESR option.
Two example printers have been added and commented on.  The first shows a virtual interface used by Papercut, and how username and password authentication can be enabled.
The second is a standard printer being added by IP address with a driver.

In addition to the above, the required lines for Firefox ESR have been added, and can be uncommented if you need it; it uses its own repository.