Update Zorin_Deploy-public.sh
Confirmed the Ubuntu specific changes work for Zorin 17.  So, if you are using this script on 16.3, don't change the Zorin 16.3 section.

If you are using Zorin 17, comment out the lines for the "Zorin 16.3 specific" section, which affects gdm.  After that, uncomment the lines for the "Ubuntu and Zorin 17 specific" section.