Create clean-accounts.sh
This is the initial version of our clean accounts script. It is dangerous if you don't know what you are doing, but it automates a potentially tedious task somewhat safely.

There is a danger of a race condition due to students logging in during the script running, so run outside of class hours.